pyproject.toml
[build-system]
requires = ["setuptools>=40.8.0", "wheel"]

[project]

name = "reduct-cli"
version = "0.2.0"
description = "CLI client for Reduct Storage"
requires-python = ">=3.7"
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["cli", "reduct-storage", "tool", "database", "time series database"]
authors = [
    { email = "atimin@gmail.com" },
    { name = "Alexey Timin" }
]
maintainers = [
    { email = "atimin@gmail.com" },
    { name = "Alexey Timin" }
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3 :: Only",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Topic :: Database :: Front-Ends"
]

dependencies = [
    "reduct-py~=1.1",
    "click~=8.1",
    "tomlkit~=0.11",
    "rich~=12.6",
]

[project.optional-dependencies]
test = [
    "pytest~=7.1",
    "pytest-mock~=3.10",
    "pytest-asyncio~=0.18"
]

lint = ["pylint~=2.14"]
format = ["black~=22.6"]
docs = [
    "mkdocs~=1.3",
    "mkdocs-material~=8.3",
    "plantuml-markdown~=3.5",
    "mkdocs-same-dir~=0.1",
    "mkdocs-click~=0.8"
]

[project.urls]
ReductStorage = "https://reduct-storage.dev"
Documentation = "https://reduct-cli.readthedocs.io/"
Source = "https://github.com/reduct-storage/reduct-cli"
Changelog = "https://github.com/reduct-storage/reduct-cli/blob/main/CHANGELOG.md"
Twitter = "https://twitter.com/ReductStorage"
Blog = "https://dev.to/reduct-storage"


[project.scripts]
reduct-cli = "reduct_cli:main"
rcli = "reduct_cli:main"

[tool.pylint]
good-names = "ls,rm"
disable = ["fixme", "duplicate-code"]
